Updates from suggestions plus tests.
